disk_block_cache: put blocks in working set if sync cache is full
Previously, if the sync cache was full, the overall cache put would
fail (without logging an error, due to the unnamed return param).
However, this means that when your synced directory is bigger than
your disk cache, performance crawls to a halt in that directory
because all new blocks need to be fetched again, even if there's room
in the working set cache.

So, if the sync cache put fails, just put it in the working set and
let eviction take care of it later.

+++ libkbfs/disk_block_cache_wrapped.go
@@ -163,8 +163,12 @@ func (cache *diskBlockCacheWrapped) Put(ctx context.Context, tlfID tlf.ID,
 	defer cache.mtx.RUnlock()
 	if cache.config.IsSyncedTlf(tlfID) && cache.syncCache != nil {
 		workingSetCache := cache.workingSetCache
-		go workingSetCache.Delete(ctx, []kbfsblock.ID{blockID})
-		return cache.syncCache.Put(ctx, tlfID, blockID, buf, serverHalf)
+		err := cache.syncCache.Put(ctx, tlfID, blockID, buf, serverHalf)
+		if err == nil {
+			go workingSetCache.Delete(ctx, []kbfsblock.ID{blockID})
+			return nil
+		}
+		// Otherwise drop through and put it into the working set cache.
 	}
 	// TODO: Allow more intelligent transitioning from the sync cache to
 	// the working set cache.
